requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
black==25.9.0
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.11.0
crispy-bootstrap5==2025.6
cron_descriptor==2.0.6
Django==5.2.7
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-filter==25.2
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-spectacular==0.28.0
drf-yasg==1.21.11
et_xmlfile==2.0.0
eventlet==0.40.3
flake8==7.3.0
forex-python==1.9.2
greenlet==3.2.4
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
simplejson==3.20.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
